Add meson options to modify Telemetry consts
Added 'min-interval', 'max-reading-parameters' and 'max-reports'
to meson_options to allow user to control Telemetry constant
values. Now user is able to adjust those values to platform.
Fixed checking maximal number of ReadingParameters in Report.
Synced type of properties with interface defined in
phosphor-dbus-interface.

Tested:
 - Unit tests passes

+++ tests/src/test_report_manager.cpp
@@ -87,7 +87,7 @@ TEST_F(TestReportManager, minInterval)
 
 TEST_F(TestReportManager, maxReports)
 {
-    EXPECT_THAT(getProperty<uint32_t>("MaxReports"),
+    EXPECT_THAT(getProperty<size_t>("MaxReports"),
                 Eq(ReportManager::maxReports));
 }
